Move game::delete_world into worldfactory
This enables proper bookkeeping about deleted worlds (they should not appear in the list of existing worlds).

It combines code from main_menu.cpp (bookeeping) and game.cpp (the actual removal).
