+++ src/worldfactory.cpp
@@ -15,6 +15,11 @@
 #include "name.h"
 #include "json.h"
 
+#if (defined _WIN32 || defined __WIN32__)
+#   include "platform_win.h"
+#   include <tchar.h>
+#endif
+
 using namespace std::placeholders;
 
 #define SAVE_MASTER "master.gsav"
@@ -1533,4 +1538,74 @@ void load_world_option( JsonObject &jo )
 mod_manager *worldfactory::get_mod_manager()
 {
     return mman.get();
+}
+
+// Helper predicate to exclude files from deletion when resetting a world directory.
+static bool isForbidden(std::string candidate)
+{
+    if (candidate.find(FILENAMES["worldoptions"]) != std::string::npos ||
+        candidate.find("mods.json") != std::string::npos) {
+        return true;
+    }
+    return false;
+}
+
+void worldfactory::delete_world( const std::string &worldname, const bool delete_folder )
+{
+    std::string worldpath = all_worlds[worldname]->world_path;
+    std::set<std::string> directory_paths;
+
+    auto file_paths = get_files_from_path("", worldpath, true, true);
+    if (!delete_folder) {
+        std::vector<std::string>::iterator forbidden = find_if(file_paths.begin(), file_paths.end(),
+                isForbidden);
+        while (forbidden != file_paths.end()) {
+            file_paths.erase(forbidden);
+            forbidden = find_if(file_paths.begin(), file_paths.end(), isForbidden);
+        }
+    }
+    for( auto &file_path : file_paths ) {
+        // strip to path and remove worldpath from it
+        std::string part = file_path.substr( worldpath.size(),
+                                             file_path.find_last_of( "/\\" ) - worldpath.size() );
+        int last_separator = part.find_last_of("/\\");
+        while (last_separator != (int)std::string::npos && part.size() > 1) {
+            directory_paths.insert(part);
+            part = part.substr(0, last_separator);
+            last_separator = part.find_last_of("/\\");
+        }
+    }
+
+#if (defined _WIN32 || defined __WIN32__)
+    for (std::vector<std::string>::iterator file = file_paths.begin();
+         file != file_paths.end(); ++file) {
+        DeleteFile(file->c_str());
+    }
+    for (std::set<std::string>::reverse_iterator it = directory_paths.rbegin();
+         it != directory_paths.rend(); ++it) {
+        RemoveDirectory(std::string(worldpath + *it).c_str());
+    }
+    if (delete_folder) {
+        RemoveDirectory(worldpath.c_str());
+    }
+#else
+    // Delete files, order doesn't matter.
+    for( auto &file_path : file_paths ) {
+        (void)remove( file_path.c_str() );
+    }
+    // Delete directories -- directories are ordered deepest to shallowest.
+    for (std::set<std::string>::reverse_iterator it = directory_paths.rbegin();
+         it != directory_paths.rend(); ++it) {
+        remove(std::string(worldpath + *it).c_str());
+    }
+    if (delete_folder) {
+        remove(worldpath.c_str());
+    }
+#endif
+
+    if( delete_folder ) {
+        remove_world( worldname );
+    } else {
+        all_worlds[worldname]->world_saves.clear();
+    }
 }
